strongly base on vultr docs
this  is the URL https://coreos.com/os/docs/latest/booting-on-vultr.html
The simplest option to boot up Container Linux is to select the "CoreOS Stable" operating system from Vultr's default offerings. However, most deployments require a custom cloud-config, which can only be achieved in Vultr with an iPXE script. The remainder of this article describes this process.